Split test module `test_anchor_actions`.
