+++ tests/simulator/main.rs
@@ -13,4 +13,7 @@ mod wrapped_appchain_token_manager;
 mod test_anchor_actions;
 mod test_beefy_light_client;
 mod test_beefy_light_client_2;
+mod test_equivocation_challenge;
+mod test_migration;
+mod test_transfer_oct_to_appchain;
 mod test_wrapped_appchain_token;
